% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA23204164.R
\name{benchmarks}
\alias{benchmarks}
\title{Benchmark R and Rcpp functions.}
\description{
Use R package \code{microbenchmark} to compare the performance of R functions (\code{BH_lmR}) and Cpp functions (\code{BH_lmC}).
}
\examples{
\dontrun{
X = matrix(rnorm(100*50),100)
nonzero = sample(50, 10)
beta =0.5 * (1:50 \%in\% nonzero)
y = X \%*\% beta + rnorm(100)
tm <- microbenchmark(
  tR = BH_lmR(X,y,0.1),
  tC = BH_lmC(X,y,0.1)
)
print(summary(tm)[,c(1,3,5,6)])
}
}
