% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BH_lmC}
\alias{BH_lmC}
\title{BH in linear regression models using Rcpp}
\usage{
BH_lmC(X, y, alpha)
}
\arguments{
\item{X}{the observed value of the predictor (numeric)}

\item{y}{the observed value of the response (numeric)}

\item{alpha}{the level of significance of variable selection (hypothesis testing)}
}
\value{
the subscript vector of the selected predictor
}
\description{
BH implementation of variable selection in linear regression models using Rcpp
}
\examples{
\dontrun{
X = matrix(rnorm(100*50),100)
nonzero = sample(50, 10)
beta = 0.5*(1:50 \%in\% nonzero)
y = X \%*\% beta + rnorm(100)
s<-BH_lmC(X,y,0.1)
}
}
