% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA23204164.R
\name{BH_lmR}
\alias{BH_lmR}
\title{BH in linear regression models using R}
\usage{
BH_lmR(X, y, alpha)
}
\arguments{
\item{X}{the observed value of the predictor (numeric)}

\item{y}{the observed value of the response (numeric)}

\item{alpha}{the level of significance of variable selection (hypothesis testing)}
}
\value{
the subscript vector of the selected predictor
}
\description{
BH implementation of variable selection in linear regression models using R
}
\examples{
\dontrun{
X = matrix(rnorm(100*50),100)
nonzero = sample(50, 10)
beta = 1*(1:50 \%in\% nonzero)
y = X \%*\% beta + rnorm(100)
s<-BH_lmR(X,y,0.1)
}
}
