% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA23204164.R
\name{FTPR}
\alias{FTPR}
\title{FPR and TPR computation using R.}
\usage{
FTPR(beta, selected)
}
\arguments{
\item{beta}{the true parameters of the predictor (numeric)}

\item{selected}{the subscript vector of the selected predictor (numeric)}
}
\value{
a list of false positive rate (FPR) and true positive rate (TPR)
}
\description{
FPR and TPR computation using R.
}
\examples{
\dontrun{
X = matrix(rnorm(100*50),100)
nonzero = sample(50, 10)
beta = 1*(1:50 \%in\% nonzero)
y = X \%*\% beta + rnorm(100)
s<-BH_lmR(X,y,0.1)
rR<-unlist(FTPR(beta,sR))
}
}
