\name{SA23204164-package}
\alias{SA23204164-package}
\docType{package}
\title{
  Final assignment for 'Statistical Computing' course.
}
\description{
  Several R and Rcpp function are included in this package for illustration, and so is all the homework.
}
\author{
Shuang Jiang

Maintainer: Shuang Jiang <2656065598@qq.com>
}
\keyword{ package }
